Move elbow coeffieicnt calc from blanket to pumping
